+++ encode.c
@@ -132,7 +132,7 @@ int main(int argc, char **argv)
 	uint8_t *line = calloc(width, channels);
 	int prev[4] = { 0, 0, 0, 0 };
 	long count = 0, limit = channels == 3 ? 255 : 4095;
-	long total = width * height;
+	long total = (long)width * (long)height;
 	for (long i = 0; i < total; ++i) {
 		int diff[3], equal = 1;
 		for (int c = 0; c < channels; ++c) {
